Make view server use ingress key data instead of ingestable ranges
The point of this is to update the block tracker so that it doesn't
mind if some blocks were never scanned, as long as we were never
obligated to scan them.

This largely rips the ingestable ranges out of the view server,
and simplifies some logic along the way. It also has to rip out
an assumption that a block always contains ETxOutRecords which
isn't necessarily true.